Desktop/AIFILEORGANISER/requirements.txt
# Private AI File Organiser - Dependencies
# Python 3.8+

# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
watchdog==3.0.0
requests==2.31.0

# Database
sqlite-utils==3.35.2

# JSON Schema Validation (for agent)
jsonschema==4.20.0

# AI/LLM
# Ollama runs as a separate service, we just need requests to communicate with it

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Security/Crypto
cryptography==41.0.7

# Optional: Enhanced text extraction
# pillow==10.1.0  # For image text extraction (OCR)
# pytesseract==0.3.10  # For OCR

# Development Dependencies (optional)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1
